+++ cpp/src/groupby/hash/output_utils.cu
@@ -100,7 +100,7 @@ struct result_column_creator {
 std::unique_ptr<table> create_results_table(size_type output_size,
                                             table_view const& values,
                                             host_span<aggregation::Kind const> agg_kinds,
-                                            host_span<int const> force_non_nullable,
+                                            host_span<int8_t const> force_non_nullable,
                                             rmm::cuda_stream_view stream,
                                             rmm::device_async_resource_ref mr)
 {
